backend/requirements.txt
# LinkifyMe Backend
# Python 3.11+

# === Core Framework ===
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# === LangGraph & LangChain ===
langgraph==0.2.0
langchain==0.2.0
langchain-openai==0.1.0

# === Google Sheets ===
google-api-python-client==2.111.0
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gspread==6.0.0

# === HTTP & Async ===
httpx==0.26.0
aiohttp==3.9.1

# === Utilities ===
python-multipart==0.0.6
orjson==3.9.10
PyYAML==6.0.1

# === Testing ===
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# === Dev Tools ===
black==24.1.0
ruff==0.1.11
